requirements.txt
beautifulsoup4
discord.py==1.7.2
requests
psycopg2-binary
cairosvg
imgkit
pytz
selenium

